requirements.txt
appnope==0.1.0
decorator==4.0.9
dj-database-url==0.4.0
Django==1.9.2
django-allauth==0.24.1
django-appconf==1.0.1
django-imagekit==3.3
gnureadline==6.3.3
ipdb==0.8.1
ipython==4.1.1
ipython-genutils==0.1.0
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
pilkit==1.1.13
Pillow==3.1.1
psycopg2==2.6.1
ptyprocess==0.5.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
wheel==0.24.0
